requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
